Add config package
Config package abstracts configuration loading so that we can load the
configuration from various sources and not just from environment
variables.
